Fix includes in the framework document
For the DE version the document was renamed to fit the capitalization
of the other documents. For the EN version the include link simply did
miss an s, the document name was already consistent with the heading.

Without this fix, the single page version of the documents, e.g.
GitHub markdown, don't contain the page's content.